+++ VideoStreaming/VideoModel.swift
@@ -1,0 +1,16 @@
+//
+//  VideoModel.swift
+//  VideoStreaming
+//
+//  Created by Natalia Brasesco on 19/09/2019.
+//  Copyright © 2019 ESPN. All rights reserved.
+//
+
+import Foundation
+
+struct Video {
+    var title: String
+    var duration: String
+    var date: Date
+    var image: String
+}
