+++ VideoStreaming/VideoPlaylistViewController.xib
@@ -9,16 +9,34 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="VideoPlaylistViewController" customModule="VideoStreaming" customModuleProvider="target">
             <connections>
+                <outlet property="tableView" destination="Vgg-ik-bmT" id="QpK-Oy-7ng"/>
                 <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
             </connections>
         </placeholder>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
         <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" id="i5M-Pr-FkT">
-            <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+            <rect key="frame" x="0.0" y="0.0" width="400" height="600"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="Vgg-ik-bmT">
+                    <rect key="frame" x="0.0" y="0.0" width="400" height="600"/>
+                    <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                    <connections>
+                        <outlet property="dataSource" destination="-1" id="Pa6-Po-qXX"/>
+                        <outlet property="delegate" destination="-1" id="pf5-b1-GHz"/>
+                    </connections>
+                </tableView>
+            </subviews>
             <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+            <constraints>
+                <constraint firstItem="Vgg-ik-bmT" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" id="5lo-nA-Spa"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="bottom" secondItem="Vgg-ik-bmT" secondAttribute="bottom" id="7To-1D-gfE"/>
+                <constraint firstItem="Vgg-ik-bmT" firstAttribute="top" secondItem="i5M-Pr-FkT" secondAttribute="top" id="Aac-eR-Omw"/>
+                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="Vgg-ik-bmT" secondAttribute="trailing" id="sPH-UF-U9d"/>
+            </constraints>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
-            <point key="canvasLocation" x="205.79710144927537" y="295.3125"/>
+            <point key="canvasLocation" x="121.73913043478262" y="34.821428571428569"/>
         </view>
     </objects>
 </document>
